ARACHNE-1980. Fix data source check function for case of DS update w/o change of password
